refactor(admin): extract ConfigTab into its own file
